Refactor QueryRepository a un QueryManager de django
