+++ api_management/apps/analytics/models.py
@@ -1,18 +1,13 @@
 import abc
 import datetime
-import hashlib
-import re
 from urllib.parse import parse_qsl, urlparse
 
-import uuid
-import requests
 from dateutil import relativedelta
-from django.conf import settings
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-from redis import Redis
 from solo.models import SingletonModel
 
+from api_management.apps.analytics.repositories.query_manager import QueryManager
 from api_management.apps.api_registry.models import KongApi
 
 
@@ -36,6 +31,8 @@ class Query(models.Model):
     request_method = models.CharField(max_length=10, null=True, default="", blank=True)
     referer = models.CharField(max_length=200, null=True, default="", blank=True)
 
+    objects = QueryManager()
+
     class Meta:  # pylint: disable=too-few-public-methods
         verbose_name = _("query")
         verbose_name_plural = _("queries")
@@ -51,88 +48,6 @@ class Query(models.Model):
 
     def api_session_id(self):
         return self.ip_address + self.api_data.name + self.user_agent
-
-
-class GoogleAnalyticsSettings(SingletonModel):
-    ga_id = models.CharField(max_length=100, blank=True, default='',
-                             verbose_name='Google analytics ID')
-
-    def __str__(self):
-        return 'GoogleAnalyticsSettings'
-
-
-def is_options_request(query):
-    return query.request_method == 'OPTIONS'
-
-
-class GoogleAnalyticsManager:
-
-    @classmethod
-    def using_settings(cls):
-        ga_id = ''
-        ga_settings = GoogleAnalyticsSettings.objects.first()
-        if ga_settings is not None:
-            ga_id = ga_settings.ga_id
-
-        return cls(tracking_id=ga_id)
-
-    def __init__(self, tracking_id):
-        self.session = requests.session()
-        self.tracking_id = tracking_id
-        self.redis_client = Redis(host=settings.REDIS_HOST,
-                                  port=settings.REDIS_PORT)
-
-    @staticmethod
-    def prepare_send_analytics(created, instance, **_):
-        if (not is_options_request(instance)) and created:
-            query = instance
-
-            GoogleAnalyticsManager.using_settings().manage_query(query)
-
-    def manage_query(self, query):
-        regex = query.api_data.kongapipluginhttplog.exclude_regex
-        if not regex \
-                or re.search(regex, query.uri) is None:
-            self.send_analytics(query)
-
-    def generate_payload(self, cid, query):
-        data = {'v': 1,  # Protocol Version
-                'cid': cid,  # Client ID
-                'tid': self.tracking_id,  # Tracking ID
-                'uip': query.ip_address,  # User IP override
-                't': 'pageview',  # Hit Type
-                'dh': query.host,  # Document HostName
-                'dp': query.uri,  # Document Path
-                'cd1': query.querystring,  # Custom Dimention
-                'cm1': query.start_time,  # Custom Metric
-                'srt': query.request_time,  # Server Response Time
-                'cm2': query.status_code,  # Custom Metric
-                'cd3': query.api_data.name,
-                'cm3': query.api_data.pk,
-                'ua': query.user_agent}
-
-        if not self.redis_client.exists(query.api_session_id()):
-            data['sc'] = 'start'  # this request starts a new session
-
-        min_to_timeout = ApiSessionSettings.get_solo().max_timeout
-        self.redis_client.append(query.api_session_id(), 'start')
-        self.redis_client.expire(query.api_session_id(), min_to_timeout*60)
-
-        return data
-
-    def send_analytics(self, query):
-        if query.token is None:
-            cid = query.ip_address + query.user_agent
-        else:
-            cid = query.token
-
-        cid = hashlib.sha1(cid.encode()).digest()
-        cid = str(uuid.UUID(bytes=cid[:16]))
-        data = self.generate_payload(cid, query)
-
-        response = self.session.post('http://www.google-analytics.com/collect', data=data)
-        if not response.ok:
-            raise ConnectionError(response.content)
 
 
 class CsvFile(models.Model):
@@ -224,13 +139,6 @@ class CsvCompressorAndRemoverTask(CsvGeneratorTaskLogger):
 
     class Meta:
         verbose_name = 'Csv remover task'
-
-
-class ApiSessionSettings(SingletonModel):
-    max_timeout = models.IntegerField(default=10, verbose_name='Timeout in minutes')
-
-    def __str__(self):
-        return 'ApiSessionSettings'
 
 
 class IndicatorMetricsRow(models.Model):
