--- api_management/apps/analytics/repositories/query_repository.py
@@ -1,29 +1,0 @@
-from api_management.apps.analytics.elastic_search.query_index import index_query
-from api_management.apps.analytics.models import Query, GoogleAnalyticsManager
-
-
-class QueryRepository:
-
-    def __init__(self, query_serializer):
-        self.query_serializer = query_serializer
-        self.query_instance = None
-
-    def save(self):
-        self._save_to_db()
-        self._index_to_es()
-        self.send_analytics()
-
-    def send_analytics(self):
-        GoogleAnalyticsManager.prepare_send_analytics(True, self.query_instance)
-
-    def _save_to_db(self):
-        self.query_serializer.is_valid(raise_exception=True)
-        instance = self.query_serializer.save()
-        self.query_instance = instance
-
-    def _index_to_es(self):
-        index_query(self.query_instance)
-
-    @staticmethod
-    def all_without_options():
-        return Query.objects.exclude(request_method='OPTIONS')
