+++ api_management/apps/analytics/test/api_tests.py
@@ -1,11 +1,10 @@
+from unittest.mock import patch
 import pytest
 from django.urls import reverse
 from faker import Faker
-from unittest.mock import patch
 from rest_framework.test import APIClient
 
 from api_management.apps.analytics.models import Query
-from api_management.apps.analytics.repositories.query_repository import QueryRepository
 from api_management.apps.analytics.test.conftest import query_dict
 from api_management.apps.analytics.test.support import query_dict_response
 
@@ -14,7 +13,8 @@ faker = Faker()  # pylint: disable=invalid-name
 
 @pytest.fixture(scope='session', autouse=True)
 def before_tests():
-    with patch.object(QueryRepository, '_index_to_es') as index_call:
+    path = 'api_management.apps.analytics.repositories.query_manager.index_query'
+    with patch(path) as index_call:
         yield index_call
         assert index_call.call_count == 5
